backend/requirements.txt
anthropic
fastapi
uvicorn
python-dotenv
pydantic
groq
openai
tavily-python
python-docx
